refactor: separated vehicles and shipment settings
