+++ packages/route-optimization-jsprit/src/main/java/com/predictivemovement/route/optimization/VRPCostMatrix.java
@@ -24,9 +24,15 @@ public class VRPCostMatrix {
     private List<String> sourceIds;
     private List<String> destinationIds;
 
+    private JSONObject routeRequest;
     private Map<String, Location> routeLocations;
 
-    public VehicleRoutingTransportCosts set(JSONObject routeRequest, Map<String, Location> locations) {
+    VRPCostMatrix(VRPSetting vrpSetting) {
+        routeRequest = vrpSetting.routeRequest;
+        routeLocations = vrpSetting.locations;
+    }
+
+    protected VehicleRoutingTransportCosts createCostMatrix() {
         if (!routeRequest.has("matrix"))
             return null;
 
@@ -35,7 +41,6 @@ public class VRPCostMatrix {
             return null;
 
         matrix = routeRequest.getJSONObject("matrix");
-        routeLocations = locations;
 
         if (matrix.has("sources")) {
             JSONArray sourcesLocation = matrix.getJSONArray("sources");
