+++ packages/route-optimization-jsprit/src/main/java/com/predictivemovement/route/optimization/VRPSettingTimeWindows.java
@@ -1,7 +1,5 @@
 package com.predictivemovement.route.optimization;
 
-import java.time.ZonedDateTime;
-import java.time.temporal.ChronoUnit;
 import java.util.function.Consumer;
 
 import com.graphhopper.jsprit.core.problem.solution.route.activity.TimeWindow;
@@ -14,41 +12,22 @@ import org.json.JSONObject;
  */
 public class VRPSettingTimeWindows {
 
-    ZonedDateTime now;
-
-    public VRPSettingTimeWindows() {
-        now = ZonedDateTime.now();
-    }
-
     public void add(JSONObject json, Consumer<TimeWindow> addTimeWindowFunction) {
 
         if (!json.has("time_windows"))
             return;
 
+        VRPSettingTimeUtils timeUtils = new VRPSettingTimeUtils();
+
         JSONArray timeWindows = json.getJSONArray("time_windows");
         for (Object window : timeWindows) {
             JSONObject jsonTimeWindow = (JSONObject) window;
 
-            double earliest = getTimeDifferenceFromNow(jsonTimeWindow, "earliest", 0.0);
-            double latest = getTimeDifferenceFromNow(jsonTimeWindow, "latest", Double.MAX_VALUE);
+            double earliest = timeUtils.getTimeDifferenceFromNow(jsonTimeWindow, "earliest", 0.0);
+            double latest = timeUtils.getTimeDifferenceFromNow(jsonTimeWindow, "latest", Double.MAX_VALUE);
 
             TimeWindow timeWindow = new TimeWindow(earliest, latest);
             addTimeWindowFunction.accept(timeWindow);
         }
-    }
-
-    private double getTimeDifferenceFromNow(JSONObject jsonTimeWindow, String field, double defaultSeconds) {
-
-        if (!jsonTimeWindow.has(field))
-            return defaultSeconds;
-
-        String dateTimeString = jsonTimeWindow.getString(field);
-        ZonedDateTime dateTime = ZonedDateTime.parse(dateTimeString);
-        double seconds = ChronoUnit.SECONDS.between(now, dateTime);
-
-        // TODO should throw an error because time window is not possible anymore?
-        seconds = seconds > 0 ? seconds : defaultSeconds;
-
-        return seconds;
     }
 }
